+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 ## Breaking changes
 * TBD
 ## Changes
-* [FIX] Fix help message of 'collector.name' flag
+* [CHANGE] Read 'location' from cmd
 
 ## 0.1.0 2019-07-18
 * [FEATURE] First version collects metrics from ds8k RESTful API
+* [FIX] Fix help message of 'collector.name' flag